managedNamespaces clarification
Clarify that if the user wants to also manage the namespace Waypoint is installed into, they must explicitly add it if managedNamespaces is set.

It might not be obvious behavior, but I think the most common case for this field is that people have an "infra" namespace with waypoint in it,
and one or more "customer" namespaces where deployments go. In that case, there's no reason waypoint should be able to access it's own "infra"
namespace, and the infra namespace may contain very sensitive things (like vault) that should be isolated.

I honestly struggle to imagine a scenario where a user would want waypoint to manage it's own namespace _and_ other namespaces, but that's
not to say the use-case doesn't exist.