Increase runner livenessprobe failure threshold
Lets say a user runs the following command:

```
 ~/dev/waypoint/waypoint runner install -platform=kubernetes -server-addr=10.0.0.10:9711 -server-tls=false
```

But `10.0.0.10` is the wrong address.

What currently happens is that the runner attempts to connect, but it's killed by kubernetes due to not passing liveness before it can report a context deadline exceeded error. It doesn't log anything - it just crashloops. This gives the user no indication as to what the problem could be.

With this change, kubernetes won't kill the pod until after the client's default connection timeout window has passed, giving the runner a chance to log the fact that it couldn't reach the server.